+++ docs/app/(home)/[[...slug]]/page.tsx
@@ -8,6 +8,8 @@ import {
 } from "fumadocs-ui/page";
 import { notFound } from "next/navigation";
 import defaultMdxComponents from "fumadocs-ui/mdx";
+import { Badge } from "@/components/ui/badge";
+import { CloudIcon } from "lucide-react";
 
 import { Tabs, Tab } from "@/components/react/tabs";
 import { Steps, Step } from "fumadocs-ui/components/steps";
@@ -21,7 +23,12 @@ import { PropertyReference } from "@/components/react/property-reference";
 import { getImageMeta } from "fumadocs-ui/og";
 import { InsecurePasswordProtected } from "@/components/react/insecure-password-protected";
 import { LinkToCopilotCloud } from "@/components/react/link-to-copilot-cloud";
-import { Accordions, Accordion} from "fumadocs-ui/components/accordion";
+import { Accordions, Accordion } from "fumadocs-ui/components/accordion";
+
+/**
+ * TODO: This should be dynamic, but it's not working.
+ */
+const cloudOnlyFeatures = ["Authenticated Actions", "Guardrails"];
 
 const mdxComponents = {
   ...defaultMdxComponents,
@@ -55,9 +62,8 @@ export default async function Page({
 }) {
   const page = source.getPage(params.slug);
   if (!page) notFound();
-
   const MDX = page.data.body;
-
+  const cloudOnly = cloudOnlyFeatures.includes(page.data.title);
   return (
     <DocsPage
       toc={page.data.toc}
@@ -69,7 +75,20 @@ export default async function Page({
         path: `/docs/content/docs/${page.file.path}`,
       }}
     >
-      <DocsTitle>{page.data.title}</DocsTitle>
+      <div className="flex items-center gap-3">
+        <DocsTitle className="flex items-center">
+          {page.data.title}
+          {cloudOnly && (
+            <Badge
+              variant="secondary"
+              className="ml-3 mt-1 inline-flex items-center gap-1.5 py-1.5 px-3 bg-indigo-600/90 text-white hover:bg-indigo-600 border-0 rounded-md transition-colors"
+            >
+              <CloudIcon className="w-3 h-3" />
+              <span className="text-xs">Cloud Only</span>
+            </Badge>
+          )}
+        </DocsTitle>
+      </div>
       <DocsDescription>{page.data.description}</DocsDescription>
       <DocsBody>
         <MDX components={mdxComponents} renderSmth={() => <div>test</div>} />